backend/requirements.txt
flask
flask_cors
firebase-admin
pymongo
dnspython
python-dotenv
nltk
textblob
